+++ packages/EasyBatch/src/Common/Repository/BatchItemRepositoryInterface.php
@@ -5,8 +5,8 @@ namespace EonX\EasyBatch\Common\Repository;
 
 use EonX\EasyBatch\Common\ValueObject\BatchCounts;
 use EonX\EasyBatch\Common\ValueObject\BatchItem;
+use EonX\EasyPagination\Pagination\PaginationInterface;
 use EonX\EasyPagination\Paginator\LengthAwarePaginatorInterface;
-use EonX\EasyPagination\ValueObject\PaginationInterface;
 
 interface BatchItemRepositoryInterface
 {
