+++ packages/EasyBatch/src/Doctrine/Repository/BatchItemRepository.php
@@ -10,9 +10,9 @@ use EonX\EasyBatch\Common\Exception\BatchItemNotFoundException;
 use EonX\EasyBatch\Common\Repository\BatchItemRepositoryInterface;
 use EonX\EasyBatch\Common\ValueObject\BatchCounts;
 use EonX\EasyBatch\Common\ValueObject\BatchItem;
+use EonX\EasyPagination\Pagination\PaginationInterface;
 use EonX\EasyPagination\Paginator\DoctrineDbalLengthAwarePaginator;
 use EonX\EasyPagination\Paginator\LengthAwarePaginatorInterface;
-use EonX\EasyPagination\ValueObject\PaginationInterface;
 
 final class BatchItemRepository extends AbstractBatchObjectRepository implements BatchItemRepositoryInterface
 {
