+++ packages/EasyPagination/laravel/EasyPaginationServiceProvider.php
@@ -5,10 +5,10 @@ namespace EonX\EasyPagination\Laravel;
 
 use EonX\EasyPagination\Laravel\Listeners\PaginationFromRequestListener;
 use EonX\EasyPagination\Laravel\Middleware\PaginationFromRequestMiddleware;
+use EonX\EasyPagination\Pagination\PaginationInterface;
 use EonX\EasyPagination\Provider\PaginationConfigProvider;
 use EonX\EasyPagination\Provider\PaginationProvider;
 use EonX\EasyPagination\Provider\PaginationProviderInterface;
-use EonX\EasyPagination\ValueObject\PaginationInterface;
 use Illuminate\Contracts\Container\Container;
 use Illuminate\Routing\Events\RouteMatched;
 use Illuminate\Support\ServiceProvider;
