+++ packages/EasyPagination/src/Paginator/AbstractPaginator.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 namespace EonX\EasyPagination\Paginator;
 
 use Closure;
-use EonX\EasyPagination\ValueObject\PaginationInterface;
+use EonX\EasyPagination\Pagination\PaginationInterface;
 
 abstract class AbstractPaginator implements PaginatorInterface
 {
