+++ packages/EasyPagination/src/Paginator/EloquentLengthAwarePaginator.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace EonX\EasyPagination\Paginator;
 
+use EonX\EasyPagination\Pagination\PaginationInterface;
 use EonX\EasyPagination\Paginator\ExtendablePaginatorInterface as ExtendableInterface;
-use EonX\EasyPagination\ValueObject\PaginationInterface;
 use Illuminate\Database\Eloquent\Model;
 
 final class EloquentLengthAwarePaginator extends AbstractLengthAwarePaginator implements ExtendableInterface
