+++ packages/EasyPagination/src/Paginator/IterablePaginator.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace EonX\EasyPagination\Paginator;
 
-use EonX\EasyPagination\ValueObject\PaginationInterface;
+use EonX\EasyPagination\Pagination\PaginationInterface;
 use EonX\EasyUtils\Common\Helper\CollectorHelper;
 
 final class IterablePaginator extends AbstractPaginator
