+++ packages/EasyPagination/src/Resolver/FromHttpFoundationRequestPaginationResolver.php
@@ -3,9 +3,9 @@ declare(strict_types=1);
 
 namespace EonX\EasyPagination\Resolver;
 
+use EonX\EasyPagination\Pagination\Pagination;
+use EonX\EasyPagination\Pagination\PaginationInterface;
 use EonX\EasyPagination\Provider\PaginationConfigProviderInterface;
-use EonX\EasyPagination\ValueObject\Pagination;
-use EonX\EasyPagination\ValueObject\PaginationInterface;
 use Symfony\Component\HttpFoundation\Request;
 
 final readonly class FromHttpFoundationRequestPaginationResolver
