+++ packages/EasyPagination/tests/Unit/src/Paginator/DoctrineOrmPaginatorTest.php
@@ -5,11 +5,11 @@ namespace EonX\EasyPagination\Tests\Unit\Paginator;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\QueryBuilder;
+use EonX\EasyPagination\Pagination\Pagination;
+use EonX\EasyPagination\Pagination\PaginationInterface;
 use EonX\EasyPagination\Paginator\DoctrineOrmPaginator;
 use EonX\EasyPagination\Tests\Stub\Entity\ChildItem;
 use EonX\EasyPagination\Tests\Stub\Entity\Item;
-use EonX\EasyPagination\ValueObject\Pagination;
-use EonX\EasyPagination\ValueObject\PaginationInterface;
 use PHPUnit\Framework\Attributes\DataProvider;
 
 final class DoctrineOrmPaginatorTest extends AbstractDoctrineOrmPaginatorTestCase
