+++ packages/EasyPagination/tests/Unit/src/Paginator/EloquentLengthAwarePaginatorTest.php
@@ -3,11 +3,11 @@ declare(strict_types=1);
 
 namespace EonX\EasyPagination\Tests\Unit\Paginator;
 
+use EonX\EasyPagination\Pagination\Pagination;
+use EonX\EasyPagination\Pagination\PaginationInterface;
 use EonX\EasyPagination\Paginator\EloquentLengthAwarePaginator;
 use EonX\EasyPagination\Tests\Stub\Model\ChildItemModel;
 use EonX\EasyPagination\Tests\Stub\Model\ItemModel;
-use EonX\EasyPagination\ValueObject\Pagination;
-use EonX\EasyPagination\ValueObject\PaginationInterface;
 use Illuminate\Database\ConnectionResolver;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
