+++ packages/EasyRepository/src/Repository/AbstractOptimizedDoctrineOrmRepository.php
@@ -11,9 +11,9 @@ use Doctrine\ORM\Exception\ORMException;
 use Doctrine\ORM\Mapping\ClassMetadata;
 use Doctrine\ORM\QueryBuilder;
 use Doctrine\Persistence\ManagerRegistry;
+use EonX\EasyPagination\Pagination\PaginationInterface;
 use EonX\EasyPagination\Paginator\DoctrineOrmLengthAwarePaginator;
 use EonX\EasyPagination\Paginator\LengthAwarePaginatorInterface;
-use EonX\EasyPagination\ValueObject\PaginationInterface;
 use EonX\EasyRepository\Exception\EasyPaginationNotInstalledException;
 use EonX\EasyRepository\Repository\PaginatedObjectRepositoryInterface as PaginatedObjRepoInterface;
 use Throwable;
