+++ packages/EasyRepository/src/Repository/PaginatedObjectRepositoryInterface.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace EonX\EasyRepository\Repository;
 
+use EonX\EasyPagination\Pagination\PaginationInterface;
 use EonX\EasyPagination\Paginator\LengthAwarePaginatorInterface;
-use EonX\EasyPagination\ValueObject\PaginationInterface;
 
 interface PaginatedObjectRepositoryInterface extends ObjectRepositoryInterface
 {
