+++ packages/EasyWebhook/src/Common/Store/SendAfterStoreInterface.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 namespace EonX\EasyWebhook\Common\Store;
 
 use DateTimeInterface;
+use EonX\EasyPagination\Pagination\PaginationInterface;
 use EonX\EasyPagination\Paginator\LengthAwarePaginatorInterface;
-use EonX\EasyPagination\ValueObject\PaginationInterface;
 
 interface SendAfterStoreInterface extends StoreInterface
 {
